ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
numpy==1.26.3
Pillow==10.2.0
opencv-python-headless==4.9.0.80
python-dotenv==1.0.1
pydantic==2.5.3
